requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
distribute==0.6.34
django-countries==2.1.2
django-dbsettings==0.6
django-pipeline==1.3.24
django-pipeline-compass==0.1.5
django-redis-sessions==0.4.0
django-tinymce==1.5.2
futures==2.1.6
httplib2==0.9
oauth2==1.5.211
pyScss==1.2.0.post3
pymongo==2.7
python-openid==2.2.5
six==1.6.1
tornado==3.2.1
wsgiref==0.1.2
python-social-auth==0.1.24
paypalrestsdk==1.1.0
gunicorn==18.0
django-mobi==0.1.7
django-widget-tweaks==1.3
django-mail-queue==2.1.2
twilio==3.6.6
django-ckeditor==4.0.2
